Nettrine implementation + minor fixes. (#3)
* Added mime types rule for attachment upload.

* Removed download functionality from TaskPresenter, added handler and callback into TasksTable component.

* Changed path separators because of Linux (-:

* Sample data updated.

* Changed task filtering parameter from constants to Enum.

* TaskStatusFilter Enum changed to TaskStatus.

* Fixed date format.

* Specified the type for $control.

* Implemented nettrine packages.

* Changed templateType to varType since TasksTemplate model is not needed anymore.

* Sample database updated.

* Implemented toArray method.

* Minor refactors.

* Fixed updateTask and removeTask methods so that they use Doctrine.

* Add a check in deleteFolder, it cannot delete anything except a subdirectory of uploads directory and its contents.

* Template is now using Entity\Task model instead of TasksTemplate.

* Sample database update.

* File model is not needed.

* Sample data update.

+++ app/Models/UploadsRepository.php
@@ -33,8 +33,17 @@ class UploadsRepository
         FileSystem::copy($file->temporaryFile, $taskFolder . DIRECTORY_SEPARATOR . $file->name);
     }
 
+    /**
+     * Removes $folder within default upload directory.
+     * @param string $folder Must not be empty and must contain alphanumeric
+     * characters only.
+     */
     public function deleteFolder(string $folder): void
     {
+        if(!ctype_alnum($folder)){
+            throw new \InvalidArgumentException(
+                "Error resolving path '$folder'.");
+        }
         FileSystem::delete($this->uploadsDir . DIRECTORY_SEPARATOR . $folder);
     }
